third_party/blink/web_tests/svg/custom/getSubStringLength-expected.txt

PASS svgText.getSubStringLength(-1, 2) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (4294967295) is greater than the maximum bound (3)..
PASS svgText.getSubStringLength(-1, 0) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (4294967295) is greater than the maximum bound (3)..
PASS svgText.getSubStringLength(1, 3) is 40
PASS svgText.getSubStringLength(0, 4) is 60
PASS svgText.getSubStringLength(3, 0) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (3) is greater than or equal to the maximum bound (3)..
PASS svgText.getSubStringLength(0, 0) is 0
PASS svgText.getSubStringLength(2, 0) is 0
PASS svgText.getSubStringLength(0, 1) is 20
PASS svgText.getSubStringLength(1, 1) is 20
PASS svgText.getSubStringLength(2, 1) is 20
PASS svgText.getSubStringLength(0, 3) is 60
PASS svgText.getSubStringLength(1, -1) did not throw exception.
PASS svgText.getSubStringLength(2, -1) did not throw exception.
PASS svgText.getSubStringLength(3, -1) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (3) is greater than or equal to the maximum bound (3)..
PASS svgText.getSubStringLength(3, -3) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (3) is greater than or equal to the maximum bound (3)..
PASS successfullyParsed is true

TEST COMPLETE
